coldata: change Batch interface to operate with int length
This commit changes Batch interface to operate with int length
because it unifies the way we store batches - now we will have a single
implementation MemBatch (with an exception for coldata.ZeroBatch). This
allows us to remove bufferedBatch.

Release note: None

+++ pkg/sql/colexec/builtin_funcs.go
@@ -58,7 +58,7 @@ func (b *defaultBuiltinFuncOperator) Next(ctx context.Context) coldata.Batch {
 	b.allocator.PerformOperation(
 		[]coldata.Vec{output},
 		func() {
-			for i := uint16(0); i < n; i++ {
+			for i := 0; i < n; i++ {
 				rowIdx := i
 				if sel != nil {
 					rowIdx = sel[i]
